Code to drive the robot from the embedded raspberry pi using ssh over wifi
The internal raspberry PI 4 8G uses a 128Gb flash card loaded with 2020-08-20-raspio-buster-full-arm.img.  The user name is qubie, and you set a system name for your local network.  These files load into the qubie user.  Turn the robot on.  On your work computer ssh to connect to the robot.  type this command but not enter.  
$ python3 robot.py
Hold the robot up manually by the body so the feet are flat on a solid surface.  Then hit return.  The robot will stand.
The face change code can run, but at this time has to be run separately afterr the robot is standing.
python 3 facechange.py